+++ Tests/SentryTests/Networking/TestSentryDispatchQueueWrapper.swift
@@ -1,7 +1,11 @@
 import Foundation
 
 class TestSentryDispatchQueueWrapper: SentryDispatchQueueWrapper {
+    
+    var dispatchAsyncCalled = 0
+    
     override func dispatchAsync(_ block: @escaping () -> Void) {
+        dispatchAsyncCalled += 1
         block()
     }