+++ scripts/ci-select-xcode.sh
@@ -3,4 +3,4 @@
 # For available Xcode versions on Github Action, see 
 # https://github.com/actions/virtual-environments/blob/main/images/macos/macos-11-Readme.md#xcode
 
-sudo xcode-select -s /Applications/Xcode_12.5.app/Contents/Developer
+sudo xcode-select -s /Applications/Xcode_12.5.1.app/Contents/Developer